Updated for proper grammar and a more natural English-like flow
Also, "functionalities" is not a word. Tried to replace it with something that conveyed the original meaning but was actual English. I'll be going through this tutorial because I'm interested in Julia, so I'll probably be trying to help edit it as I go through the pages. Its a really cool and well-made tutorial and only suffers from bad writing, which I'll do my best to help with! (: